+++ src/js/control-bar/progress-control/progress-control.js
@@ -3,6 +3,7 @@
  */
 import Component from '../../component.js';
 import SeekBar from './seek-bar.js';
+import MouseTimeDisplay from './mouse-time-display.js';
 
 /**
  * The Progress Control component contains the seek bar, load progress,