+++ src/js/control-bar/progress-control/seek-bar.js
@@ -7,6 +7,7 @@ import LoadProgressBar from './load-progress-bar.js';
 import PlayProgressBar from './play-progress-bar.js';
 import * as Fn from '../../utils/fn.js';
 import formatTime from '../../utils/format-time.js';
+import assign from 'object.assign';
 
 /**
  * Seek Bar and holder for the progress bars
@@ -126,6 +127,7 @@ class SeekBar extends Slider {
 SeekBar.prototype.options_ = {
   children: {
     'loadProgressBar': {},
+    'mouseTimeDisplay': {},
     'playProgressBar': {}
   },
   'barName': 'playProgressBar'