+++ CHANGELOG.md
@@ -130,6 +130,7 @@ CHANGELOG
 * @heff improved css selector strengths ([view](https://github.com/videojs/video.js/pull/2583))
 * @heff moved scss vars to be private ([view](https://github.com/videojs/video.js/pull/2584))
 * @heff added a fancy loading spinner ([view](https://github.com/videojs/video.js/pull/2582))
+* @gkatsev added a mouse-hover time display to the progress bar ([view](https://github.com/videojs/video.js/pull/2569))
 
 --------------------